requirements.txt
google-genai
Flask
Flask-Cors